backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.28.0
pandas==1.3.3
modelscope==1.9.5
scikit-learn==0.24.2
numpy==1.21.2
chromadb==0.4.22
langchain==0.1.0
sentence-transformers==2.2.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
transformers==4.11.3 